requirements.txt
absl-py==0.12.0
blessings==1.7
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
clingo==5.5.0.post3
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
distro==1.5.0
google-auth==1.30.2
google-auth-oauthlib==0.4.4
gpustat==0.6.0
grpcio==1.38.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
install==1.3.4
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
packaging==20.9
pandas==1.1.5
Pillow==8.2.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
rtpt==0.0.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
setproctitle==1.2.2
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
tifffile==2020.9.3
#torch==1.8.1+cu111
#torchaudio==0.8.1
torchsummary==1.5.1
#torchvision==0.9.1+cu111
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
Werkzeug==2.0.1
zipp==3.4.1
pathos==0.2.8
